requirements.txt
Django==5.2.8
django-crispy-forms==2.3
crispy-bootstrap5==2025.6
Pillow==11.0.0
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-decouple==3.8
django-filter==24.3